Rest com spring

+++ algafood/src/main/java/com/algafood/infrastructure/repository/CidadeRepositoryImpl.java
@@ -5,9 +5,11 @@ import java.util.List;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
+import com.algafood.domain.exception.EntidadeNaoEncontradaException;
 import com.algafood.domain.model.Cidade;
 import com.algafood.domain.repository.CidadeRepository;
 
@@ -35,8 +37,13 @@ public class CidadeRepositoryImpl implements CidadeRepository {
 
 	@Transactional
 	@Override
-	public void remover(Cidade cidade) {
-		cidade = buscar(cidade.getId());
+	public void remover(Long cidadeId) {
+		Cidade cidade = buscar(cidadeId);
+		
+		if(cidade == null) {
+			throw new EmptyResultDataAccessException(1);
+		}
+		
 		entityManager.remove(cidade);
 	}
 }
