+++ algafood/src/main/java/com/algafood/infrastructure/repository/EstadoRepositoryImpl.java
@@ -6,6 +6,7 @@ import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 import javax.persistence.TypedQuery;
 
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -37,8 +38,13 @@ public class EstadoRepositoryImpl implements EstadoRepository {
 
 	@Transactional
 	@Override
-	public void remover(Estado estado) {
-		estado = buscar(estado.getId());
+	public void remover(Long estadoId) {
+		Estado estado = buscar(estadoId);
+		
+		if(estado == null) {
+			throw new EmptyResultDataAccessException(1);
+		}
+		
 		entityManager.remove(estado);
 	}
