+++ algafood/src/main/java/com/algafood/infrastructure/repository/RestauranteRepositoryImpl.java
@@ -6,6 +6,7 @@ import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 import javax.persistence.TypedQuery;
 
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -39,8 +40,12 @@ public class RestauranteRepositoryImpl implements RestauranteRepository {
 
 	@Override
 	@Transactional
-	public void remover(Restaurante restaurante) {
-		restaurante = buscar(restaurante.getId());
+	public void remover(Long restauranteId) {
+		Restaurante restaurante = buscar(restauranteId);
+		
+		if(restaurante == null) 
+			throw new EmptyResultDataAccessException(1);
+		
 		entityManager.remove(restaurante);
 	}	
 }
